# openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Etuna API
  description: |
    Hotel Etuna PMS (Property Management System) API
    
    A comprehensive hospitality management platform with:
    - Multi-tenant architecture (Hub-and-Spoke model)
    - Booking and folio management
    - Restaurant and inventory operations
    - Sofia AI concierge
    - Compliance (PSD-12, AML, KYC)
    - Fraud detection
    - Payment processing (Adumo Virtual, NamQR)
    
    ## Authentication
    Most endpoints require authentication via Stack Auth. Use `withApiAuth` wrapper.
    
    ## Rate Limiting
    Authenticated endpoints have rate limiting enabled.
    
    ## Multi-Tenancy
    All tenant-scoped data is isolated via RLS policies.
  version: 1.0.0
  contact:
    name: Hotel Etuna Support
    email: admin@hoteletuna.com
servers:
  - url: https://hoteletuna.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: List bookings for a property
      description: Returns bookings for a specific property with optional date filtering
      security:
        - cookieAuth: []
      parameters:
        - name: propertyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Property UUID
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Start date filter (YYYY-MM-DD)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Bookings
      summary: Create a new booking
      description: Creates a booking for an authenticated user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/sofia/email:
    post:
      tags:
        - Sofia AI
      summary: Send email via Sofia
      description: Send an email using the Sofia email service
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSend'
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: stack-hs
      description: Stack Auth session cookie
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: CRON_SECRET for scheduled jobs
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingReference:
          type: string
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: string
          format: date
        numGuests:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, checked_in, checked_out, cancelled]
        totalAmount:
          type: number
          format: decimal
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        guestId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    BookingCreate:
      type: object
      required: [checkInDate, checkOutDate, numGuests, roomId]
      properties:
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: string
          format: date
        numGuests:
          type: integer
          minimum: 1
        roomId:
          type: string
          format: uuid
...
    EmailSend:
      type: object
      required: [to, subject, emailBody]
      properties:
        to:
          type: string
          format: email
        subject:
          type: string
        emailBody:
          type: string
        ctaLink:
          type: string
          format: uri
        ctaText:
          type: string
        templateId:
          type: string
        metadata:
          type: object
    EmailResult:
      type: object
      properties:
        message:
          type: string
        messageId:
          type: string

    # CRM schemas
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string